feat(MFA): Restructures for reset 2FA flow
