+++ src/enums/contact-mfa-type.ts
@@ -1,0 +1,8 @@
+/**
+ * Contact multi factor authentication type
+ * TODO: Move to common
+ */
+export enum CONTACT_MFA_TYPE {
+  TOTP = 'totp',
+  // E.g. U2F, EMAIL, SMS, HOTP, etc.
+}
