+++ src/enums/reset-security-flow-type.ts
@@ -1,0 +1,8 @@
+/**
+ * Reset security flow type enum.
+ * TODO: Move to common
+ */
+export enum RESET_SECURITY_FLOW_TYPE {
+  PASSWORD = 'password',
+  TOTP = 'totp', // See CONTACT_MFA_TYPE.TOTP
+}
