+++ src/pages/admin/contacts/[id]/index.vue
@@ -259,7 +259,7 @@ import {
 import { formatLocale } from '@utils/dates';
 import { fetchContent } from '@utils/api/content';
 import { fetchContactMfa, deleteContactMfa } from '@utils/api/contact-mfa';
-import { ContactMfaType } from '@utils/api/api.interface';
+import { CONTACT_MFA_TYPE } from '@enums/contact-mfa-type';
 import { fetchCallout, fetchResponses } from '@utils/api/callout';
 
 import { addNotification } from '@store/notifications';
@@ -306,7 +306,7 @@ const disableMfaAndNotify = async () => {
 const disableMfa = async () => {
   try {
     await deleteContactMfa(props.contact.id, {
-      type: ContactMfaType.TOTP,
+      type: CONTACT_MFA_TYPE.TOTP,
     });
   } catch (error) {
     onDeleteMfaError();
@@ -372,7 +372,7 @@ onBeforeMount(async () => {
 
   // Fetch MFA information
   const contactMfa = await fetchContactMfa(props.contact.id);
-  if (contactMfa && contactMfa.type === ContactMfaType.TOTP) {
+  if (contactMfa && contactMfa.type === CONTACT_MFA_TYPE.TOTP) {
     mfa.value.isEnabled = true;
   }
